refactor: enum parse methods to one file
