+++ transforms/utils/enumValues.js
@@ -1,0 +1,9 @@
+
+const addEnumValues = (values = []) => {
+  if (values.length === 0) {
+    return {};
+  }
+  return { enum: values };
+};
+
+module.exports = addEnumValues;
