+++ transforms/utils/formatDescription.js
@@ -1,0 +1,19 @@
+
+const ENUM_IDENTIFIER = 'enum:';
+const DESCRIPTION_DIVIDER = ' - ';
+
+const formatDescription = description => {
+  const descriptionTypes = (description || '').split(DESCRIPTION_DIVIDER);
+  const descriptionValue = descriptionTypes.filter(value => (
+    !value.includes(ENUM_IDENTIFIER)
+  )).join(DESCRIPTION_DIVIDER);
+  const enumOption = descriptionTypes.find(value => value.includes(ENUM_IDENTIFIER));
+  if (!enumOption) {
+    return [descriptionValue];
+  }
+  const [, enumOptions] = enumOption.split('enum:');
+  const enumValues = enumOptions.split(',');
+  return [descriptionValue, enumValues];
+};
+
+module.exports = formatDescription;
